refactor prefect docs mcp: extract turbopuffer logic to _internal
create `_internal/` directory to organize implementation details away from main server code:
- move `metrics.py` to `_internal/metrics.py`
- extract turbopuffer transformations to `_internal/tpuf.py`:
  - `row_to_dict` (normalize turbopuffer rows to dicts)
  - `normalize_score` (coerce scores to floats)
  - `run_query` (query with automatic retry on schema errors)

clean up `server.py` from 194 → 137 lines, keeping only tool logic and error handling. turbopuffer-specific implementation details now isolated in dedicated module.
