+++ examples/prefect_docs_mcp/src/prefect_docs_mcp/_internal/__init__.py
@@ -1,0 +1,1 @@
+"""internal utilities for the prefect docs mcp server."""
