+++ examples/prefect_docs_mcp/src/prefect_docs_mcp/server.py
@@ -2,7 +2,6 @@
 
 from __future__ import annotations
 
-import json
 from typing import Annotated, Any
 
 from fastmcp import FastMCP
@@ -11,58 +10,22 @@ from pydantic import Field
 from turbopuffer import (
     APIError,
     AuthenticationError,
-    BadRequestError,
     NotFoundError,
     PermissionDeniedError,
 )
-from turbopuffer.types.namespace_query_response import Row as TurboPufferRow
 
-from prefect_docs_mcp.metrics import metrics
+from prefect_docs_mcp._internal.metrics import metrics
+from prefect_docs_mcp._internal.tpuf import normalize_score, row_to_dict, run_query
 from prefect_docs_mcp.settings import settings
 from raggy.vectorstores.tpuf import TurboPuffer
 
-prefect_docs_mcp = FastMCP(
-    "Prefect Docs MCP",
-    instructions=(
-        "Expose the Prefect documentation via a single semantic search tool backed by Raggy."
-    ),
-)
-
-
-def _row_to_dict(row: TurboPufferRow | dict[str, Any]) -> dict[str, Any]:
-    """Normalize TurboPuffer row objects to plain dictionaries."""
-
-    if isinstance(row, dict):
-        data = dict(row)
-    else:
-        data = row.model_dump(mode="python")
-        if row.model_extra:
-            data.update(row.model_extra)
-
-    metadata = data.get("metadata")
-    if isinstance(metadata, str):
-        try:
-            data["metadata"] = json.loads(metadata)
-        except json.JSONDecodeError:
-            pass
-
-    return data
-
-
-def _normalize_score(value: Any) -> float | None:
-    """Coerce TurboPuffer scores into floats when present."""
-
-    if value is None:
-        return None
-    try:
-        return float(value)
-    except (TypeError, ValueError):
-        return None
+prefect_docs_mcp = FastMCP("Prefect Docs MCP", version="0.1.0")
 
 
 def _build_response(
     query: str, results: list[dict[str, Any]], error: str | None = None
 ) -> dict[str, Any]:
+    """build a response payload for the search tool."""
     payload: dict[str, Any] = {
         "query": query,
         "namespace": settings.namespace,
@@ -73,34 +36,13 @@ def _build_response(
     return payload
 
 
-def _run_query(
-    tpuf: TurboPuffer,
-    *,
-    query: str,
-    top_k: int,
-    include_attributes: list[str],
-):
-    include_arg = include_attributes or None
-    try:
-        return tpuf.query(
-            text=query,
-            top_k=top_k,
-            include_attributes=include_arg,
-        )
-    except BadRequestError as exc:
-        if include_arg is not None and "include_attributes" in str(exc).lower():
-            # Retry without extra attribute hints; schema likely lacks one of them.
-            return tpuf.query(text=query, top_k=top_k)
-        raise
-
-
-@prefect_docs_mcp.tool(name="SearchPrefect")
+@prefect_docs_mcp.tool(name="search_prefect_docs")
 def search_prefect(
     query: Annotated[
         str,
         Field(
             description=(
-                "A natural language query to search against Prefect documentation and examples."
+                "a search query to find relevant document excerpts from Prefect's knowledgebase"
             )
         ),
     ],
@@ -109,11 +51,11 @@ def search_prefect(
         Field(
             ge=1,
             le=20,
-            description="Override for the number of search results to return.",
+            description="How many document excerpts to return.",
         ),
     ] = None,
 ) -> dict[str, Any]:
-    """Search the Prefect documentation vector store for relevant passages."""
+    """Search the Prefect knowledgebase for documentation on concepts, usage, and best practices."""
 
     if not query.strip():
         raise ValueError("Query must not be empty.")
@@ -127,7 +69,7 @@ def search_prefect(
         try:
             with TurboPuffer(namespace=settings.namespace) as tpuf:
                 with metrics.timer("search_vector_query"):
-                    response = _run_query(
+                    response = run_query(
                         tpuf,
                         query=query,
                         top_k=result_limit,
@@ -146,9 +88,6 @@ def search_prefect(
         except OpenAIError as exc:
             metrics.increment("search_calls_error")
             return _build_response(query, [], f"OpenAI error: {exc}")
-        except BadRequestError as exc:
-            metrics.increment("search_calls_error")
-            return _build_response(query, [], f"TurboPuffer query error: {exc}")
         except ValueError:
             metrics.increment("search_calls_error")
             return _build_response(
@@ -162,7 +101,7 @@ def search_prefect(
         results: list[dict[str, Any]] = []
         with metrics.timer("search_response_format"):
             for row in rows:
-                data = _row_to_dict(row)
+                data = row_to_dict(row)
                 snippet = data.get("text") or data.get("content")
                 # no slicing - content is already chunked at ingestion time
                 snippet_text = str(snippet) if snippet else ""
@@ -170,7 +109,7 @@ def search_prefect(
                 result_payload: dict[str, Any] = {
                     "snippet": snippet_text,
                     "id": data.get("id"),
-                    "score": _normalize_score(
+                    "score": normalize_score(
                         data.get("score")
                         or data.get("distance")
                         or data.get("similarity")
